feat: Updated Desertion and Morale Behavior - Added Marriage Morale Modifier - Adjusted guidelines for item theft on desertion - Simplified code for identifying spare parts - Improved handling of Unit theft - Modified the user interface for implementing changes in morale rules.
