+++ MekHQ/resources/mekhq/resources/CampaignOptionsDialog.properties
@@ -340,7 +340,7 @@ lblTheftValue.toolTipText=The average percentage of unit funds stolen during a h
 chkUseTheftParts.text=Enable Theft of Parts
 chkUseTheftParts.toolTipText=Deserting personnel may steal parts on their way out.
 lblTheftPartsDiceCount.text=Part Theft Dice Count
-lblTheftPartsDiceCount.toolTipText=<html>How many d6 (or batches of 3d6 armor or ammo) should be rolled to determine the number of parts stolen.<br>Set to 0 to only have one item (or 3d6 units of ammunition/armor) stolen per theft.</html>
+lblTheftPartsDiceCount.toolTipText=<html>How many d3 items (or batches of 2d6 armor plates or ammo rounds) should be stolen per theft.<br>Set to 0 to only have one item (or 2d6 units of ammunition/armor) stolen per theft.</html>
 
 turnoverAndRetentionMoraleModifiersPanel.title=Morale Modifiers
 lblCustomMoraleModifier.text=Custom Modifier
@@ -359,6 +359,8 @@ chkUseMoraleModifierLoyalty.text=Personal Loyalty
 chkUseMoraleModifierLoyalty.toolTipText=Personal loyalty adjusts both desertion and mutiny rolls. Stacks with the Loyalty force reliability method.
 chkUseMoraleModifierCabinFever.text=Cabin Fever
 chkUseMoraleModifierCabinFever.toolTipText=Desertions are disabled while in transit. To compensate, this options adds a -1 modifier to Mutiny rolls.
+chkUseMoraleModifierMarriage.text=Marriage
+chkUseMoraleModifierMarriage.toolTipText=Married personal apply a +1 bonus to desertion rolls, but will desert as a couple. Mutinies are unaffected.
 lblMoraleModifierLeadershipMethod.text=Leadership Style
 lblMoraleModifierLeadershipMethod.toolTipText=What style of leadership best describes the unit?
