+++ MekHQ/resources/mekhq/resources/Morale.properties
@@ -13,6 +13,10 @@ moraleReportLow.text=Caution, morale is %s.
 moraleReportMutiny.text=Your personnel are actively planning a coup.
 moraleReportRecovered.text=Your personnel are placated, for now.
 
+# Reclaim Original Unit
+reclaimSuccessful.text=has taken back %s
+reclaimFailed.text=was unable to reclaim their original unit.
+
 # Desertion
 desertionAwolExtended.text=has extended their unauthorized leave.
