+++ MekHQ/src/mekhq/campaign/personnel/turnoverAndRetention/Morale/Mutiny.java
@@ -1,0 +1,4 @@
+package mekhq.campaign.personnel.turnoverAndRetention.Morale;
+
+public class Mutiny {
+}
